Rename flavor_name_check.py to flavor_names.py
