+++ Tests/iaas/flavor-naming/flavor_names.py
